Moved `simulator_flags` code to `simulator_config_functions`.
