+++ systems/analysis/test/simulator_config_functions_test.cc
@@ -2,15 +2,44 @@
 
 #include <gtest/gtest.h>
 
+#include "drake/common/nice_type_name.h"
 #include "drake/common/yaml/yaml_read_archive.h"
 #include "drake/common/yaml/yaml_write_archive.h"
+#include "drake/systems/analysis/simulator.h"
 #include "drake/systems/framework/leaf_system.h"
+#include "drake/systems/primitives/constant_vector_source.h"
+
+using drake::yaml::YamlReadArchive;
+using drake::yaml::YamlWriteArchive;
 
 namespace drake {
 namespace systems {
 namespace {
 
-// A hollow shell of a System.
+GTEST_TEST(SimulatorFlagsTest, ResetIntegrator) {
+  ConstantVectorSource<double> source(2);
+  Simulator<double> simulator(source);
+  const void* prior_integrator = &simulator.get_integrator();
+  IntegratorBase<double>& result = ResetIntegratorFromFlags(
+      &simulator, "runge_kutta2", 0.001);
+  EXPECT_NE(&simulator.get_integrator(), prior_integrator);
+  EXPECT_EQ(&simulator.get_integrator(), &result);
+  EXPECT_EQ(NiceTypeName::Get(result),
+            "drake::systems::RungeKutta2Integrator<double>");
+}
+
+GTEST_TEST(SimulatorFlagsTest, GetSchemes) {
+  const std::vector<std::string>& schemes = GetIntegrationSchemes();
+  EXPECT_GE(schemes.size(), 5);
+
+  // Check that all of the schemes are actually valid.
+  ConstantVectorSource<double> source(2);
+  Simulator<double> simulator(source);
+  for (const auto& one_scheme : schemes) {
+    ResetIntegratorFromFlags(&simulator, one_scheme, 0.001);
+  }
+}
+
 class DummySystem final : public drake::systems::LeafSystem<double> {
  public:
   DummySystem() {}
@@ -18,7 +47,7 @@ class DummySystem final : public drake::systems::LeafSystem<double> {
 
 GTEST_TEST(SimulatorConfigFunctionsTest, ExtractDefaultsTest) {
   const DummySystem dummy;
-  drake::systems::Simulator<double> simulator(dummy);
+  Simulator<double> simulator(dummy);
 
   // We can always extract the defaults.
   const SimulatorConfig defaults = ExtractSimulatorConfig(simulator);
@@ -43,14 +72,14 @@ GTEST_TEST(SimulatorConfigFunctionsTest, RoundTripTest) {
   const YAML::Node bespoke_yaml = YAML::Load(bespoke_str);
 
   // Ensure that the string and the struct have the same fields.
-  drake::yaml::YamlReadArchive::Options options;
+  YamlReadArchive::Options options;
   options.allow_cpp_with_no_yaml = false;
   options.allow_yaml_with_no_cpp = false;
   SimulatorConfig bespoke;
-  drake::yaml::YamlReadArchive(bespoke_yaml, options).Accept(&bespoke);
+  YamlReadArchive(bespoke_yaml, options).Accept(&bespoke);
 
   // Ensure that a round trip through the archive process does not corrupt data.
-  drake::yaml::YamlWriteArchive archive;
+  YamlWriteArchive archive;
   archive.Accept(bespoke);
   const std::string readback_str = archive.EmitString("");
   EXPECT_EQ(bespoke_str, readback_str);
@@ -58,7 +87,7 @@ GTEST_TEST(SimulatorConfigFunctionsTest, RoundTripTest) {
   // Ensure that a roundtrip through the Accept and Extract functions does not
   // corrupt data.
   const DummySystem dummy;
-  drake::systems::Simulator<double> simulator(dummy);
+  Simulator<double> simulator(dummy);
   ApplySimulatorConfig(&simulator, bespoke);
   const SimulatorConfig readback = ExtractSimulatorConfig(simulator);
   EXPECT_EQ(readback.integration_scheme, bespoke.integration_scheme);
